[fixup] address comments
