+++ src/OrbitQt/HistogramWidget.h
@@ -21,6 +21,13 @@
 #include "Statistics/Histogram.h"
 
 namespace orbit_histogram_widget {
+
+// This method returns a vector specifying the width to be drawn for each bin of a histogram
+// Takes two positive intergers, returns a vector `result` of non-negative integers s.t.
+// their sum equals to `histogram_width` and for all `i` and `j` `max(result[i] - result[j]) <= 1`
+// `histogram_width` represents the width of the histogarm in pixels.
+[[nodiscard]] std::vector<int> GenerageBinWidth(size_t number_of_bins, int histogram_width);
+
 // Implements a widget that draws a histogram.
 // If the histogram is empty, draws a textual suggestion to select a function.
 class HistogramWidget : public QWidget {
